unify eth address parsing
