+++ engine/cli/Cargo.toml
@@ -14,6 +14,7 @@ serde = {version = "1.0", features = ["derive", "rc"]}
 slog = {version = "2.7.0", features = ["max_level_trace", "release_max_level_trace"]}
 structopt = "0.3.23"
 tokio = {version = "1.5.0", features = ["full"]}
+utilities = {path = "../../utilities"}
 
 # State chain
 pallet-cf-rewards = {path = "../../state-chain/pallets/cf-rewards"}
